backend/requirements.txt
# HealthSync AI - Backend Dependencies
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
# psycopg2-binary==2.9.9

# MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Supabase
# supabase==2.3.0
# gotrue==2.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
langgraph
langchain
langchain-groq
python-multipart==0.0.6
cryptography==41.0.8

# AI/ML Services
groq==0.4.1
replicate==0.22.0
openai==1.6.1
google-generativeai==0.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
httpx==0.26.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.4
scipy==1.11.4

# ML & Data Science
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
matplotlib==3.8.2

# Image Processing & Computer Vision
Pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.9
scikit-image==0.21.0
imageio==2.31.5

# Advanced AI/ML Models
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
onnxruntime==1.16.3
ultralytics==8.0.196

# OCR & Text Processing
easyocr==1.7.0
paddleocr==2.7.0
pytesseract==0.3.10
spacy==3.7.2

# Medical AI & NLP
datasets==2.14.6
accelerate==0.24.1
sentencepiece==0.1.99

# RAG System Dependencies
sentence-transformers==2.7.0
faiss-cpu==1.13.1

# Image Enhancement & Analysis
albumentations==1.3.1
segment-anything==1.0
timm==0.9.12

# Caching & Performance
redis==5.0.1
diskcache==5.6.3
joblib==1.3.2

# API Integrations
huggingface-hub==0.19.4
roboflow==1.1.9

# WebSocket & Real-time
websockets==12.0
python-socketio==5.10.0

# HTTP & API
aiofiles==23.2.1
python-dateutil==2.8.2
email-validator==2.1.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2

# Rate Limiting
slowapi==0.1.9

# Email Services
resend==0.6.0

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
pytest-mock==3.12.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0

# Production
gunicorn==21.2.0
prometheus-client==0.19.0